added custom tindercard components
