--- stinder/App.tsx
@@ -1,42 +1,0 @@
-import React from 'react';
-import {Text, Image, View, StyleSheet} from 'react-native';
-
-
-const App = () => {
-  return (
-    <View style={styles.pageContainer}>
-      <View style={styles.card}>
-        <Image source={{ uri: 'https://notjustdev-dummy.s3.us-east-2.amazonaws.com/avatars/elon.png'}} style={styles.image} 
-        />
-      </View>
-    </View>
-  );
-};
-
-const styles = StyleSheet.create({
-  pageContainer: {
-    justifyContent: 'center', 
-    alignItems: 'center', 
-    flex: 1,
-  },
-  card: {
-    width: '95%',
-    height: '70%',
-    shadowColor: "#000",
-    shadowOffset: {
-    width: 1,
-    height: 6,
-    },
-    shadowOpacity: 1,
-    shadowRadius: 7.49,
-
-    elevation: 12,
-  },
-  image: {
-    width: '100%', 
-    height: '90%',
-    borderRadius: 10,
-  },
-});
-
-export default App;
